% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{append_capabilities}
\alias{append_capabilities}
\title{append_capabilities}
\usage{
append_capabilities(x, resolution, data_dir)
}
\arguments{
\item{x}{dataset to append capabilties to}

\item{resolution}{if "weekly" then monthly capabilities will be converted to epiweek before join}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
append EIA summer and winter capabilities to HydroSource
}
