% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correct_lat_lon_for_MOSART_stream_network}
\alias{correct_lat_lon_for_MOSART_stream_network}
\title{correct_lat_lon_for_MOSART_stream_network}
\usage{
correct_lat_lon_for_MOSART_stream_network(x)
}
\value{
time series with violations and other plant detail
}
\description{
Uses updated lat/lon data (provided by Tian Zhou) to correct snapped lat-lon; ensures correct positioning of RoR dams on stream network
}
