% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{dl_usace_generation}
\alias{dl_usace_generation}
\title{dl_usace_generation}
\usage{
dl_usace_generation(save_to = "library")
}
\arguments{
\item{save_to}{character where to save data ("library" will update package library, NULL will assume package editing)}
}
\value{
two csvs of p_min p_max p_mean for monthly and weekly periods
}
\description{
downloads hourly data from USACE Data Query and performs mean, min, max summarizing before updating relevant package files
}
