% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_month_to_epiweek}
\alias{gen_month_to_epiweek}
\title{gen_month_to_epiweek}
\usage{
gen_month_to_epiweek()
}
\value{
table of month to epiweek conversion
}
\description{
convert month to epiweek (approximate)
}
