% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameters.R
\name{generate_WECC_parameters}
\alias{generate_WECC_parameters}
\title{generate_WECC_parameters}
\usage{
generate_WECC_parameters(
  year = 2009,
  resolution = "monthly",
  data_dir,
  WM_case = "/WM_dev_base_case_cropped/"
)
}
\arguments{
\item{year}{year to be generated}

\item{resolution}{resolution of output ("weekly", "monthly")}

\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{WM_case}{path to WM simulation results}
}
\value{
tibble of pmean, pmax, pmin for 317 WECC dams
}
\description{
gets p_mean, p_max, p_min for given water input, defined by year
}
