% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax_pmin.R
\name{get_pmax_pmin_params}
\alias{get_pmax_pmin_params}
\title{get_pmax_pmin_params}
\usage{
get_pmax_pmin_params(zone, mode = "monthly", data_dir, smooth_params = TRUE)
}
\arguments{
\item{zone}{CRB or WECC}

\item{mode}{resolution of output for CRB}

\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{smooth_params}{should weekly parameters (CRB only) be smoothed?}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
gets pmax and pmin estimates for all dams
}
