% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax_pmin.R
\name{get_pmax_pmin_predictions}
\alias{get_pmax_pmin_predictions}
\title{get_pmax_pmin_predictions}
\usage{
get_pmax_pmin_predictions()
}
\value{

}
\description{
gets a prediction of pmax and pmin parameters using categorical variable in HydroSource
}
