% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{get_pmean}
\alias{get_pmean}
\title{get_pmean}
\usage{
get_pmean(
  pcm = "gridview",
  NERC = "WECC",
  data_dir,
  WM_case = "/WM_dev_base_case_cropped/",
  mode = "monthly",
  hyd_year = 2009
)
}
\arguments{
\item{pcm}{e.g. "gridview", "aurora", ...}

\item{NERC}{North American Electric Reliability Corporation region (e.g., WECC)}

\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{WM_case}{directory containing WM netcdfs (e.g., "WM_dev_base_case/")}

\item{hyd_year}{hydrological year of pmean output}

\item{output}{three options: "weekly" (weekly resolution pmean), "monthly" (monthly resolution pmean), or "evaluation" (returns observed monthly generation and predicted monthly generation for all plants)}

\item{models}{the output of get_pmean_models()}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
gets pmean for all hydro plants
}
