% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{get_pmean_models}
\alias{get_pmean_models}
\title{get_pmean_models}
\usage{
get_pmean_models(
  pcm = "none",
  NERC = NULL,
  data_dir,
  WM_case = "/WM_dev_base_case_cropped/"
)
}
\arguments{
\item{pcm}{e.g. "gridview", "plexos", ...leave as "none" for all CONUS plants, unformatted}

\item{NERC}{North American Electric Reliability Corporation region (e.g., WECC)}

\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{WM_case}{directory containing WM netcdfs (e.g., "WM_dev_base_case_cropped/")}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
gets pmean for all hydro plants
}
