% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_EIA_capabilities}
\alias{read_EIA_capabilities}
\title{read_EIA_capabilities}
\usage{
read_EIA_capabilities(data_dir)
}
\arguments{
\item{data_dir}{data directory containing hydrofixr consolidated inputs}
}
\description{
reads EIA plant capabilities
}
