% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_EIA_netgen}
\alias{read_EIA_netgen}
\title{read_EIA_netgen}
\usage{
read_EIA_netgen(data_dir)
}
\arguments{
\item{data_dir}{data directory containing hydrofixr consolidated inputs}
}
\description{
read_EIA_netgen
}
