% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_gridview_plant_data}
\alias{read_gridview_plant_data}
\title{read_gridview_plant_data}
\usage{
read_gridview_plant_data()
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
gets gridview names and joins to EIA IDs for clean matching to
}
