% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_HydroSource}
\alias{read_HydroSource}
\title{read_HydroSource}
\usage{
read_HydroSource(data_dir, NERC = NULL)
}
\arguments{
\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{NERC}{North American Electric Reliability Corporation region (e.g., WECC)}
}
\value{
tibble of hydrosource plant data, filtered for chosen NERC
}
\description{
reads existing hydropower assets data from HydroSource
}
