% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_monthly_p_and_water}
\alias{read_monthly_p_and_water}
\title{read_monthly_p_and_water}
\usage{
read_monthly_p_and_water(data_dir, region)
}
\arguments{
\item{data_dir}{data directory containing hydrofixr consolidated inputs}

\item{region}{"UC" (upper Colorado), "Col" (Columbia / PNW), "WECC"}
}
\description{
gets pmax, pmin, and water availability (observed)
}
